Change to cancancan as cancan is no longer supported
This led to some issues when downloading spreadsheets so I had to make a few more updates to make sure the spreadsheets can be downloaded.
I was able to work out what caused the issue but so far have been unable to determine why it causes it (file and commit which introduced the issue https://github.com/CanCanCommunity/cancancan/pull/495/files#diff-bc7f860bbee1c5954a600d27386afe3aa1abb5e305c7ae36c5b1f838f1f313dd).

While doing this I found that an admin file could be downloaded without the restriction that the user needed to be an admin and so I resolved that too.