Add ability to change when a framework is live through the database
Similar to work done in Facilities Management, I’ve created a table that has the service frameworks hosted on the legacy system.
It is identical to FM except there is an additional `service` attribute to distinguish between the legacy services.

I then created the pages so an admin user in the lower environments can change them if necessary.
However, this data does not do anything to the services yet but is prerequisite for future work.