As we are going to be making some major changes to some of these services in the future, I thought it was about time we added some feature tests.
I added the config for cucumber (borrowed from FM) and then added some basic feature tests for the start pages of the three pages. These included:
- Start page tests
- Navigation link tests
- Sign up validation tests
- Accessibility tests

I also added the config to run some of these tests in the CI.

The test suite will also be expanded in the future for each service.