I have done the work to shift the namespace of Management Consultancy so that it has the framework as part of the URL. Even though there is not going to be a new framework for MC, this is required as it makes the changes to LS and ST easier in the long run.
To make this changes I undertook the following steps:
- Renamed the MC tables to include the RM6187 framework
- Rename the model and service files to include the framework and update their specs
- Update the routes and the controllers to have RM6187 as a namespace
- Added the additional logic to handle the cases where an incorrect framework or no framework is provided

This has clearly resulted in a great number of changes but I made sure we had a thorough unit and feature test suite before making these changes. This means that I am confident with this commit that Management Consultancy will continue to work as expected and will not affect our users in a detrimental way.