I have done the work to shift the namespace of Legal Services so that it has the framework as part of the URL.
To make this changes I undertook the following steps:
- Renamed the LS tables to include the RM3788 framework
- Rename the model and service files to include the framework and update their specs
- Update the routes and the controllers to have RM3788 as a namespace
- Added the additional logic to handle the cases where an incorrect framework or no framework is provided

This has clearly resulted in a great number of changes but I made sure we had a thorough unit and feature test suite before making these changes. This means that I am confident with this commit that Legal Services will continue to work as expected and will not affect our users in a detrimental way.