In this commit I have added the log in for the buyer section of Supply Teachers. This includes the DFE log in.
While I have not been able to verify if it works properly, I have done some local testing which has been successful. We won’t be 100% sure it will work until it is in the preview environment.

In addition I have done a fairly major refactor of the authentication and registration sections of the application. We had a lot of repeated code and logic and I have been able to remove most of it. This will make adding new frameworks/services much easier in the future.