Due to an issue with the legacy release 2.5.3, the database migration did not work due to models being removed. This has meant the admin sections for MC, LS and ST are not working.
I have had to comment out the original migration as that will no longer work both due to the model being removed and an update to rails. I have then added a new migration which will fix the issue with the LS upload for real this time.

I have tested it locally and it works, we will not be able to test this in the lower environments.