I’ve updated how our assets work, both CSS and JavaScript.
For CSS I’ve made use of [CSS Bundling for Rails](https://github.com/rails/cssbundling-rails) with Dart Sass to compile our css (this did not require much change).

More importantly, I’ve migrated from Webpack and Shakapacker to use Rollup to compile our Javascript. Our JavaScript setup is not that complicated and configuring and updating webpacker has been a
pain point. Rollup is more simple to configure, is faster and the only downside is that we no longer get automatic page reloads when we make changes.

I’ve also reorganised the tests so that we lint before running all the major tests (which should speed up CI and reduce the cost).